<refactor>[PayDev]: <Change file structure to abstract classes>
