+++ paydev/src/main/java/dev/paydev/paydev/controllers/TransactionController.java
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RestController;
 
 import dev.paydev.paydev.domain.ViewModels.TransactionViewModel;
 import dev.paydev.paydev.domain.transaction.Transaction;
-import dev.paydev.paydev.repository.services.TransactionService;
+import dev.paydev.paydev.services.contract.ITransactionService;
 
 
 @RestController
 @RequestMapping("/api/v1/transaction")
 public class TransactionController {
-    private final TransactionService _transactionService;
+    private final ITransactionService _transactionService;
 
-    public TransactionController(TransactionService transactionService) {
+    public TransactionController(ITransactionService transactionService) {
         this._transactionService = transactionService;
     }
