+++ paydev/src/main/java/dev/paydev/paydev/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package dev.paydev.paydev.repository.contract;
+package dev.paydev.paydev.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
