+++ paydev/src/main/java/dev/paydev/paydev/services/business/BalanceService.java
@@ -1,4 +1,4 @@
-package dev.paydev.paydev.repository.services;
+package dev.paydev.paydev.services.business;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -6,11 +6,12 @@ import java.util.Map;
 import org.springframework.stereotype.Service;
 
 import dev.paydev.paydev.domain.user.User;
-import dev.paydev.paydev.repository.contract.UserRepository;
+import dev.paydev.paydev.repository.UserRepository;
+import dev.paydev.paydev.services.contract.IBalanceService;
 import dev.paydev.paydev.utils.exception.ResourceNotFoundException;
 
 @Service
-public class BalanceService {
+public class BalanceService implements IBalanceService {
     private final UserRepository _userRepository;
 
     public BalanceService(UserRepository userRepository) {
