+++ paydev/src/main/java/dev/paydev/paydev/services/contract/IBalanceService.java
@@ -1,0 +1,11 @@
+package dev.paydev.paydev.services.contract;
+
+import java.util.Map;
+
+import dev.paydev.paydev.domain.user.User;
+import dev.paydev.paydev.repository.UserRepository;
+
+public interface IBalanceService {
+    public Map<String, Object> GetUserBalance(Long id);
+    public User updateUserBalance(Long userId, double ammount);
+}
